Improve code related to API call
