+++ src/components/Main/index.js
@@ -3,13 +3,12 @@ import './styles.scss'
 import { TiTimes } from 'react-icons/ti'
 import { FiClock, FiChevronLeft, FiChevronRight } from 'react-icons/fi'
 
-import axios from 'axios'
+
 import Popover from '../Popover'
 
-// import magnifyingGlass from '../../images/searchbar_icon.svg'
+import { getLocalTime } from '../../utils/dateUtil'
 
-import capitalize from '../../utils/capitalize'
-import { setRoundDate } from '../../utils/dateUtil'
+import fetch from '../../services/api'
 
 export default () => {
     // const [term, setTerm] = useState('')
@@ -25,12 +24,7 @@ export default () => {
     const [weather, setWeather] = useState({})
 
     const setClock = timezone => {
-        const date = new Date()
-        const localTime = date.getTime()
-        const localOffset = date.getTimezoneOffset() * 60000
-        const utc = localTime + localOffset
-        let fetchedCity = utc + (timezone * 1000)
-        let time = (new Date(fetchedCity)).toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit' })
+        let time = getLocalTime(timezone).toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit' })
         setCurrentTime(time)
     }
 
@@ -51,68 +45,14 @@ export default () => {
     })
     // const [backgroundClassName, setBackgoundClassName] = useState('')
 
-    const getWeekday = (date, { lang, abbreviated }) => new Intl.DateTimeFormat(lang, { weekday: abbreviated ? 'short' : 'long' }).format(date)
+
 
     const getWeather = async place => {
-        const language = 'en-US'
-        const res = await axios.get(`https://api.openweathermap.org/data/2.5/onecall?lat=${place.lat}&lon=${place.lon}&exclude=minutely&units=metric&appid=${process.env.REACT_APP_API_ID}`)
-        console.log(res.data)
-        const currentDate = new Date()
-        currentDate.setMinutes(0, 0, 0)
-        const tomorrowDate = setRoundDate(currentDate)
-        const nextDayDate = setRoundDate(tomorrowDate)
-        const sunrise = new Date(res.data.current.sunrise * 1000)
-        const sunset = new Date(res.data.current.sunset * 1000)
-        const weatherObject = {
-            timezoneOffset: res.data.timezone_offset,
-            today: {
-                day: currentDate.toLocaleDateString(language, { timeZone: res.data.timezone, weekday: 'long' }),
-                dateString: currentDate.toLocaleDateString(language, { timeZone: res.data.timezone }),
-                temperature: Math.floor(res.data.current.temp),
-                weather: {
-                    code: `${(new Date()) >= sunrise && (new Date()) <= sunset ? 'day' : 'night'}-${res.data.current.weather[0].id}`,
-                    desc: capitalize(res.data.current.weather[0].description)
-                },
-                hourly: res.data.hourly.filter(item => {
-                    let itemDate = new Date(item.dt * 1000)
-                    if (itemDate > currentDate && itemDate < tomorrowDate) return item
-                }).map(item => {
-                    let itemHour = new Date(item.dt * 1000)
-                    const obj = {}
-                    obj.hour = itemHour.toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit' })
-                    obj.weather = `${itemHour >= sunrise && itemHour <= sunset ? 'day' : 'night'}-${item.weather[0].id}`
-                    obj.temperature = Math.floor(item.temp)
-                    return obj
-                })
-            },
-            tomorrow: res.data.hourly.filter(item => {
-                let itemDate = new Date(item.dt * 1000)
-                if (itemDate >= tomorrowDate && itemDate < nextDayDate) return item
-            }).map(item => {
-                let tomorrowSunrise = new Date(sunrise)
-                tomorrowSunrise.setDate(tomorrowSunrise.getDate() + 1)
-                let tomorrowSunset = new Date(sunset)
-                tomorrowSunset.setDate(tomorrowSunset.getDate() + 1)
-                let itemHour = new Date(item.dt * 1000)
-                const obj = {}
-                obj.hour = itemHour.toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit' })
-                obj.weather = `${itemHour >= tomorrowSunrise && itemHour <= tomorrowSunset ? 'day' : 'night'}-${item.weather[0].id}`
-                obj.temperature = Math.floor(item.temp)
-                return obj
-            }),
-            daily: res.data.daily.filter((item, i) => i > 0 && i < 6).map(item => {
-                const date = new Date(item.dt * 1000)
-                return {
-                    day: getWeekday(date, { lang: language }),
-                    weather: `${(new Date()) >= sunrise && (new Date()) <= sunset ? 'day' : 'night'}-${item.weather[0].id}`,
-                    temperature: Math.floor(item.temp.day)
-                }
-            })
-        }
-        setWeather(weatherObject)
-        setClock(weatherObject.timezoneOffset)
-        // console.log(weatherObject)
-        setReel({ ...reel, mode: 'today', items: [...weatherObject.today.hourly] })
+        const response = await fetch(place)
+        setWeather(response)
+        setClock(response.timezoneOffset)
+        // console.log(response)
+        setReel({ ...reel, mode: 'today', items: [...response.today.hourly] })
         setCurrentPlace({ city: place.city, country: place.country })
     }
