+++ src/components/Popover/index.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useContext } from 'react'
+import React, { useState, useEffect } from 'react'
 import './styles.scss'
 import { HiSearch } from 'react-icons/hi'
