+++ src/utils/dateUtil.js
@@ -5,4 +5,30 @@ const setRoundDate = date => {
     return newRoundDate
 }
 
-export { setRoundDate }
+const getLocalTime = (timezone, timestamp) => {
+    const date = timestamp ? new Date(timestamp * 1000) : new Date()
+    const localTime = date.getTime()
+    const localOffset = date.getTimezoneOffset() * 60000
+    const utc = localTime + localOffset
+    const fetchedCity = utc + (timezone * 1000)
+    return new Date(fetchedCity)
+}
+
+class LocalTime {
+    constructor(object) {
+        this.timezone = object.timezone_offset
+        this.sunrise = object.current.sunrise
+        this.sunset = object.current.sunset
+    }
+    getTime() {
+        return getLocalTime(this.timezone)
+    }
+    getSunrise() {
+        return getLocalTime(this.timezone, this.sunrise)
+    }
+    getSunset() {
+        return getLocalTime(this.timezone, this.sunset)
+    }
+}
+
+export { setRoundDate, getLocalTime, LocalTime }
